Change references to Galileo to Guru Maps
It's the same app, just they had to change the
name. Here's their explanation for what happened
(from a Facebook post):

We renamed the Galileo and that's why:

The Galileo has acquired a fame of a reliable and
stable powerful map app for the past 9 years and
we love that name.

Two years ago, the European Union registered the
Galileo trademark for its global navigation
satellite system:
https://euipo.europa.eu/eSearch/….

Now, due to trademark infringement, we were forced
to change the name of the application. The icon
will remain the same, so that the application is
easier to find.

We wanted to find a new name that would be easy to
remember and match the existing product. Please
love and favor - Guru Maps.